- Section 6 Code files

+++ app/src/main/java/com/example/jonnd/fuelfinder/database/UserRepository.java
@@ -75,7 +75,7 @@ public class UserRepository implements UserDAO {
     }
 
     @Override
-    public LiveData<User> loadUserWithId(int id) {
+    public LiveData<User> loadUserWithId(long id) {
         return mDao.loadUserWithId(id);
     }
